# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_proarash_ecommerce-server_internal_media.Media:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      file_name:
        type: string
      file_size:
        type: integer
      id:
        type: integer
      media_type:
        type: string
      mime_type:
        type: string
      updatedAt:
        type: string
      url:
        type: string
    type: object
...
  github_com_proarash_ecommerce-server_internal_staff.StaffUser:
    properties:
      avatar_media:
        $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_media.Media'
      avatar_media_id:
        type: integer
      createdAt:
        type: string
      default_avatar_id:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      mobile:
        type: string
      name:
        type: string
      role:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
      wallet:
        $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_wallet.Wallet'
    type: object
  github_com_proarash_ecommerce-server_internal_types.ApiResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  github_com_proarash_ecommerce-server_internal_types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  github_com_proarash_ecommerce-server_internal_types.PaginatedResponse:
    properties:
      items: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_proarash_ecommerce-server_internal_wallet.Wallet:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      locked:
        type: boolean
      owner_id:
        type: integer
      owner_type:
        enum:
        - user
        - staff
        type: string
      updatedAt:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_admin.CreateStaffDto:
    properties:
      avatar_media_id:
        minimum: 1
        type: integer
      default_avatar_id:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        example: 1
        maximum: 5
        minimum: 1
        type: integer
      mobile:
        example: "09121112233"
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 72
        minLength: 6
        type: string
      role:
        enum:
        - storekeeper
        - accountant
        - marketer
        - support
        type: string
    required:
    - mobile
    - name
    - password
    - role
    type: object
  internal_admin.StatsResponse:
    properties:
      blog_posts:
        type: integer
      open_chats:
        type: integer
      orders:
        type: integer
      paid_orders:
        type: integer
      products:
        type: integer
      revenue:
        type: number
      staff:
        type: integer
      users:
        type: integer
    type: object
  internal_admin.UpdateStaffStatusDto:
    properties:
      status:
        type: boolean
    required:
    - status
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: 'E-commerce backend: catalog, cart, orders, Zibal payments, inventory,
    CMS, support chat and notifications. All JSON responses are wrapped in {"message":
    string, "data": any}.'
  title: E-Commerce Server API
  version: "1.0"
paths:
...
  /admin/staff:
    get:
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Role
        enum:
        - admin
        - storekeeper
        - accountant
        - marketer
        - support
        in: query
        name: role
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.PaginatedResponse'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_staff.StaffUser'
                        type: array
                    type: object
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: List staff users
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Creates storekeeper, accountant, marketer or support users. Creating
        another admin is not allowed. default_avatar_id is a preset frontend avatar
        index (1..5); avatar_media_id references an uploaded MinIO media asset.
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Staff user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_admin.CreateStaffDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_staff.StaffUser'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create staff user
      tags:
      - Admin
  /admin/staff/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Staff ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_admin.UpdateStaffStatusDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_staff.StaffUser'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Enable or disable staff user
      tags:
      - Admin
  /admin/stats:
    get:
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_admin.StatsResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: System statistics
      tags:
      - Admin
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"